# backend/src/api/routes/boq.py
"""
BoQ Management API routes
"""
from typing import Optional
from uuid import UUID
from fastapi import APIRouter, Depends, UploadFile, File, Form, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession

from src.core.database import get_db
from src.core.security import get_current_user, estimator_required
from src.schemas.boq import BoQUploadResponse, BoQPreview
from src.schemas.common import JobStatus, PaginatedResponse
from src.models.boq_item import BoQItem
from src.models.chapter import Chapter

router = APIRouter()


@router.post("/upload", response_model=BoQUploadResponse)
async def upload_boq(
    file: UploadFile = File(...),
    project_id: Optional[UUID] = Form(None),
    create_new_project: bool = Form(False),
    project_name: Optional[str] = Form(None),
    db: AsyncSession = Depends(get_db),
    current_user = Depends(estimator_required),
):
    """
    Upload and parse BoQ Excel file
    """
    # Validate file extension
    allowed_extensions = [".xlsx", ".xls"]
    if not any(file.filename.endswith(ext) for ext in allowed_extensions):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Invalid file format. Only .xlsx and .xls files are allowed."
        )
    
    # TODO: Implement actual Excel parsing and processing
    # For now, return a mock response
    return BoQUploadResponse(
        job_id=UUID("00000000-0000-0000-0000-000000000001"),
        status="completed",
        preview=BoQPreview(
            project_id=UUID("00000000-0000-0000-0000-000000000001"),
            total_items=50,
            total_chapters=3,
            chapters=[],
            warnings=[]
        ),
        message="BoQ uploaded and parsed successfully"
    )


@router.get("/{job_id}/preview", response_model=BoQPreview)
async def get_boq_preview(
    job_id: UUID,
    db: AsyncSession = Depends(get_db),
    current_user = Depends(get_current_user),
):
    """
    Get parsed BoQ preview
    """
    # TODO: Implement actual preview retrieval
    return BoQPreview(
        project_id=UUID("00000000-0000-0000-0000-000000000001"),
        total_items=50,
        total_chapters=3,
        chapters=[],
        warnings=[]
    )


@router.get("/{job_id}/items", response_model=PaginatedResponse)
async def list_boq_items(
    job_id: UUID,
    chapter_id: Optional[UUID] = None,
    status: Optional[str] = None,
    page: int = 1,
    page_size: int = 50,
    db: AsyncSession = Depends(get_db),
    current_user = Depends(get_current_user),
):
    """
    List BoQ items with pagination
    """
    # TODO: Implement actual item listing
    return PaginatedResponse(
        items=[],
        total=0,
        page=page,
        page_size=page_size,
        total_pages=0
    )