# specs/001-boq-price-forecast/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: BoQ Price Forecast API
  version: 1.0.0
  description: |
    REST API for AI-powered BoQ price adjustment and forecasting.
    Supports Excel upload, weight attribution, price recalculation, forecasting, and defense document generation.
  contact:
    name: BoQ Price Forecast Team
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.boqprice.example.com
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
        params:
          type: object
      required: [detail]
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
        result:
          type: object
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # ===== BoQ =====
    BoQUploadResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        preview:
          $ref: '#/components/schemas/BoQPreview'
        message:
          type: string
    BoQPreview:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        total_items:
          type: integer
        total_chapters:
          type: integer
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterPreview'
        warnings:
          type: array
          items:
            type: string
    ChapterPreview:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
        code:
          type: string
          pattern: '^\d{3}$'
        name:
          type: string
        item_count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/BoQItemPreview'
    BoQItemPreview:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
        code:
          type: string
        description_fa:
          type: string
        unit:
          type: string
        base_price:
          type: number
          format: double
        quantity:
          type: number
          format: double
        row_number:
          type: integer

    # ===== Weight Attribution =====
...
    MarketDataFreshness:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/MarketDataSourceStatus'
        last_updated:
          type: string
          format: date-time
        staleness_warnings:
          type: array
          items:
            type: string
    MarketDataSourceStatus:
      type: object
      properties:
        source_id:
          type: string
          format: uuid
        name:
          type: string
        component_code:
          type: string
        last_run_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [success, failed, stale, unknown]
        records_count:
          type: integer
        error_message:
          type: string
    ETLTriggerRequest:
      type: object
      properties:
        source_ids:
          type: array
          items:
            type: string
            format: uuid
        force:
          type: boolean
          default: false
      required: [source_ids]

    # ===== Administration =====
    ModelRetrainRequest:
      type: object
      properties:
        component_codes:
          type: array
          items:
            type: string
        model_types:
          type: array
          items:
            type: string
            enum: [prophet, arima, lstm]
        full_retrain:
          type: boolean
          default: false
    ModelRetrainResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sso_id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [estimator, reviewer, manager, admin]
        department:
          type: string
        is_active:
          type: boolean
        last_login:
          type: string
          format: date-time
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        client_name:
          type: string
        tender_date:
          type: string
          format: date
        base_date:
          type: string
          format: date
        risk_buffer:
          type: number
          format: double
        payment_terms:
          type: number
          format: double
        profit_margin:
          type: number
          format: double
        fusion_alpha:
          type: number
          format: double
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        client_name:
          type: string
        tender_date:
          type: string
          format: date
        base_date:
          type: string
          format: date
        risk_buffer:
          type: number
          format: double
          default: 1.04
        payment_terms:
          type: number
          format: double
          default: 1.08
        profit_margin:
          type: number
          format: double
          default: 1.10
        fusion_alpha:
          type: number
          format: double
          default: 0.70
      required: [name, base_date]
paths:
...
  /boq/upload:
    post:
      tags: [BoQ Management]
      summary: Upload and parse BoQ Excel file
      operationId: uploadBoQ
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                project_id:
                  type: string
                  format: uuid
                create_new_project:
                  type: boolean
                  default: false
                project_name:
                  type: string
              required: [file]
      responses:
        '200':
          description: Upload successful, parsing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoQUploadResponse'
        '400':
          description: Invalid file format or parsing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /boq/{job_id}/preview:
    get:
      tags: [BoQ Management]
      summary: Get parsed BoQ preview
      operationId: getBoQPreview
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: BoQ preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoQPreview'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /boq/{job_id}/items:
    get:
      tags: [BoQ Management]
      summary: List BoQ items with pagination
      operationId: listBoQItems
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: chapter_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: Paginated list of BoQ items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  # ===== Weight Attribution =====
...
  /market-data/freshness:
    get:
      tags: [Market Data]
      summary: Get market data freshness status
      operationId: getMarketDataFreshness
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Market data freshness
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDataFreshness'
  /market-data/sources:
    get:
      tags: [Market Data]
      summary: List market data sources
      operationId: listMarketDataSources
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of market data sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketDataSourceStatus'
  /market-data/etl/trigger:
    post:
      tags: [Market Data]
      summary: Manually trigger ETL for sources
      operationId: triggerETL
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ETLTriggerRequest'
      responses:
        '202':
          description: ETL triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
  /market-data/etl/logs:
    get:
      tags: [Market Data]
      summary: Get ETL run logs
      operationId: getETLLogs
      security:
        - BearerAuth: []
      parameters:
        - name: source_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: ETL run logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        source_name:
                          type: string
                        started_at:
                          type: string
                          format: date-time
                        completed_at:
                          type: string
                          format: date-time
                        status:
                          type: string
                        records_processed:
                          type: integer
                        error_message:
                          type: string

  # ===== Administration =====
  /admin/models/retrain:
    post:
      tags: [Administration]
      summary: Trigger forecast model retraining
      operationId: retrainModels
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRetrainRequest'
      responses:
        '202':
          description: Retraining job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRetrainResponse'
  /admin/models:
    get:
      tags: [Administration]
      summary: List forecast models
      operationId: listModels
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        component_code:
                          type: string
                        model_type:
                          type: string
                        version:
                          type: integer
                        status:
                          type: string
                        mae:
                          type: number
                        trained_at:
                          type: string
                          format: date-time
  /admin/users:
    get:
      tags: [Administration]
      summary: List users
      operationId: listUsers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
  /admin/projects:
    get:
      tags: [Administration]
      summary: List projects
      operationId: listProjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectResponse'

    post:
      tags: [Administration]
      summary: Create new project
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /admin/projects/{project_id}:
    get:
      tags: [Administration]
      summary: Get project details
      operationId: getProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Administration]
      summary: Update project settings
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                risk_buffer:
                  type: number
                  format: double
                payment_terms:
                  type: number
                  format: double
                profit_margin:
                  type: number
                  format: double
                fusion_alpha:
                  type: number
                  format: double
                status:
                  type: string
                  enum: [draft, active, archived]
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'

  # ===== Health =====
...